Updated Square, Piece, and Game class
Progress:
So far, the board is drawn based on canvas size, the pieces are colored and put on their starting squares when the game starts. The pieces can be pressed and dragged to be moved, and they lock into the square they are released on. 

Stuff to be done:
We need to make a boolean method "isValidMove" to only let the piece be moved to a square if that is legal, figure out how to remove a piece, figure out how to keep the turn going if additional jumps are able, and make a method that finds if there are any valid moves, and if not ends the game.